requirements.txt
requests
python-roborock
mijiaAPI
pyshark